+++ src/Symfony/Bridge/ProxyManager/Tests/LazyProxy/PhpDumper/ProxyDumperTest.php
@@ -60,6 +60,27 @@ class ProxyDumperTest extends \PHPUnit_Framework_TestCase
         );
     }
 
+    public function testGetProxyFactoryCodeWithCustomMethod()
+    {
+        $definition = new Definition(__CLASS__);
+
+        $definition->setLazy(true);
+
+        $code = $this->dumper->getProxyFactoryCode($definition, 'foo', 'getFoo2Service');
+
+        $this->assertStringMatchesFormat(
+            '%wif ($lazyLoad) {%wreturn $this->services[\'foo\'] =%s'
+            .'SymfonyBridgeProxyManagerTestsLazyProxyPhpDumperProxyDumperTest_%s(%wfunction '
+            .'(&$wrappedInstance, \ProxyManager\Proxy\LazyLoadingInterface $proxy) {'
+            .'%w$wrappedInstance = $this->getFoo2Service(false);%w$proxy->setProxyInitializer(null);'
+            .'%wreturn true;%w}%w);%w}%w',
+           $code
+        );
+    }
+
+    /**
+     * @group legacy
+     */
     public function testGetProxyFactoryCode()
     {
         $definition = new Definition(__CLASS__);