+++ src/Symfony/Component/DependencyInjection/Tests/Loader/XmlFileLoaderTest.php
@@ -519,12 +519,18 @@ class XmlFileLoaderTest extends \PHPUnit_Framework_TestCase
 
         $deprecations = array();
         set_error_handler(function ($type, $msg) use (&$deprecations) {
-            if (E_USER_DEPRECATED === $type) {
-                $deprecations[] = $msg;
+            if (E_USER_DEPRECATED !== $type) {
+                restore_error_handler();
+
+                return call_user_func_array('PHPUnit_Util_ErrorHandler::handleError', func_get_args());
             }
+
+            $deprecations[] = $msg;
         });
 
         $loader->load('legacy_invalid_alias_definition.xml');
+
+        restore_error_handler();
 
         $this->assertTrue($container->has('bar'));
 
@@ -532,7 +538,5 @@ class XmlFileLoaderTest extends \PHPUnit_Framework_TestCase
         $this->assertContains('Using the attribute "class" is deprecated for alias definition "bar"', $deprecations[0]);
         $this->assertContains('Using the element "tag" is deprecated for alias definition "bar"', $deprecations[1]);
         $this->assertContains('Using the element "factory" is deprecated for alias definition "bar"', $deprecations[2]);
-
-        restore_error_handler();
     }
 }