+++ src/Symfony/Component/Form/Extension/Core/Type/DateTimeType.php
@@ -203,7 +203,7 @@ class DateTimeType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $compound = function (Options $options) {
-            return $options['widget'] !== 'single_text';
+            return 'single_text' !== $options['widget'];
         };
 
         // Defaults to the value of "widget"