+++ src/Symfony/Component/Process/Pipes/UnixPipes.php
@@ -150,17 +150,4 @@ class UnixPipes extends AbstractPipes
     {
         return (bool) $this->pipes;
     }
-
-    /**
-     * Creates a new UnixPipes instance.
-     *
-     * @param Process         $process
-     * @param string|resource $input
-     *
-     * @return UnixPipes
-     */
-    public static function create(Process $process, $input)
-    {
-        return new static($process->isTty(), $process->isPty(), $input, !$process->isOutputDisabled() || $process->hasCallback());
-    }
 }