+++ src/Symfony/Component/Validator/Constraints/Callback.php
@@ -38,7 +38,7 @@ class Callback extends Constraint
             $options = $options['value'];
         }
 
-        if (is_array($options) && !isset($options['callback']) && !isset($options['groups'])) {
+        if (is_array($options) && !isset($options['callback']) && !isset($options['groups']) && !isset($options['payload'])) {
             $options = array('callback' => $options);
         }