+++ src/Symfony/Component/VarDumper/Tests/Caster/SplCasterTest.php
@@ -57,12 +57,6 @@ SplFileInfo {
   pathname: "https://google.com/about"
   extension: ""
   realPath: false
-  writable: false
-  readable: false
-  executable: false
-  file: false
-  dir: false
-  link: false
 %A}
 EOTXT
             ),