+++ src/Symfony/Component/Yaml/Tests/DumperTest.php
@@ -277,20 +277,18 @@ EOF;
         );
     }
 
-    public function testBinaryDataIsDumpedAsIsWithoutFlag()
-    {
-        $binaryData = file_get_contents(__DIR__.'/Fixtures/arrow.gif');
-        $expected = "{ data: '".str_replace("'", "''", $binaryData)."' }";
-
-        $this->assertSame($expected, $this->dumper->dump(array('data' => $binaryData)));
-    }
-
-    public function testBinaryDataIsDumpedBase64EncodedWithFlag()
+    public function testBinaryDataIsDumpedBase64Encoded()
     {
         $binaryData = file_get_contents(__DIR__.'/Fixtures/arrow.gif');
         $expected = '{ data: !!binary '.base64_encode($binaryData).' }';
 
-        $this->assertSame($expected, $this->dumper->dump(array('data' => $binaryData), 0, 0, Yaml::DUMP_BASE64_BINARY_DATA));
+        $this->assertSame($expected, $this->dumper->dump(array('data' => $binaryData)));
+    }
+
+    public function testNonUtf8DataIsDumpedBase64Encoded()
+    {
+        // "für" (ISO-8859-1 encoded)
+        $this->assertSame('!!binary ZsM/cg==', $this->dumper->dump("f\xc3\x3fr"));
     }
 
     /**