+++ src/Symfony/Component/Yaml/Yaml.php
@@ -26,9 +26,8 @@ class Yaml
     const PARSE_OBJECT_FOR_MAP = 8;
     const DUMP_EXCEPTION_ON_INVALID_TYPE = 16;
     const PARSE_DATETIME = 32;
-    const DUMP_BASE64_BINARY_DATA = 64;
-    const DUMP_OBJECT_AS_MAP = 128;
-    const DUMP_MULTI_LINE_LITERAL_BLOCK = 256;
+    const DUMP_OBJECT_AS_MAP = 64;
+    const DUMP_MULTI_LINE_LITERAL_BLOCK = 128;
 
     /**
      * Parses YAML into a PHP value.