+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+3.1.0
+-----
+
+ * Deprecated the backup feature of the file dumper classes.
+
 3.0.0
 -----