+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -4,13 +4,13 @@ CHANGELOG
 3.1.0
 -----
 
+ * Strings that are not UTF-8 encoded will be dumped as base64 encoded binary
+   data.
+
  * Added support for dumping multi line strings as literal blocks.
 
  * Added support for parsing base64 encoded binary data when they are tagged
    with the `!!binary` tag.
-
- * Added support for dumping binary data as base64 encoded strings by passing
-   the `Yaml::DUMP_BASE64_BINARY_DATA` flag.
 
  * Added support for parsing timestamps as `\DateTime` objects: